+++ saleor/plugins/webhook/tests/subscription_webhooks/test_create_deliveries_for_subscription.py
@@ -106,6 +106,36 @@ def test_account_confirmation_requested(
     assert deliveries[0].webhook == webhooks[0]
 
 
+def test_account_change_email_requested(
+    customer_user, channel_USD, subscription_account_change_email_requested_webhook
+):
+    # given
+    webhooks = [subscription_account_change_email_requested_webhook]
+    event_type = WebhookEventAsyncType.ACCOUNT_CHANGE_EMAIL_REQUESTED
+    new_email = "new@example.com"
+
+    # when
+    deliveries = create_deliveries_for_subscriptions(
+        event_type,
+        {
+            "user": customer_user,
+            "channel_slug": channel_USD.slug,
+            "token": "token",
+            "new_email": new_email,
+            "redirect_url": "http://www.mirumee.com?token=token",
+        },
+        webhooks,
+    )
+
+    # then
+    expected_payload = generate_account_events_payload(
+        customer_user, channel_USD, new_email=new_email
+    )
+    assert deliveries[0].payload.payload == expected_payload
+    assert len(deliveries) == len(webhooks)
+    assert deliveries[0].webhook == webhooks[0]
+
+
 def test_account_delete_requested(
     customer_user, channel_USD, subscription_account_delete_requested_webhook
 ):